# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic Auth",
        "title": "Basic Auth",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/recvMessage": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Receive a message with the specified number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Receive message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The number parameter",
                        "name": "number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sendMessage": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "send message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "send message",
                "parameters": [
                    {
                        "description": "Input message and number",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.MessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.MessageRequest": {
            "description": "The message request body",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}